backend/requirements.txt
aerich==0.7.2
aioboto3==13.1.1
aiofiles==24.1.0
aiohttp==3.10.11
asyncpg==0.30.0
dataclasses-json==0.6.7
fastapi==0.115.7
injectq[fastapi,taskiq]==0.3.3
firebase-admin==6.5.0
httpx==0.27.0
orjson==3.11.3
phonenumbers==8.13.44
pydantic==2.12.0
pydantic-settings==2.11.0
prometheus-fastapi-instrumentator==7.1.0
python-docx==1.1.2
python-multipart==0.0.20
python-dateutil==2.9.0
redis==5.3.1
sentry-sdk==2.10.0
snowflakekit==0.1.1
taskiq==0.11.13
taskiq-dependencies==1.5.7
taskiq-redis==1.0.2
pytest-asyncio==0.25.3
pytest-env==1.2.0
tortoise==0.1.1
tortoise-orm==0.25.1
uvicorn==0.30.1
